Remove stale attachment requests from the JSON DB.
Do not error out if the attachment request database gets severely out of
sync with the actual StorPool cluster contents.  Until now, a stale
attachment request in the JSON file for a volume that no longer existed
in the StorPool cluster would block any and all attempts to attach or
detach further StorPool volumes; the OpenStack installation was not
a happy bunny.

ObQuote:	"Everything you know is wrong, just forget the words and sing along"