Do not clean up unknown StorPool volume attachments.
While this might be a good idea in general, it causes problems in setups
where OpenStack services are running in more than one
host/machine/container with the same StorPool ID (SP_OURID), e.g. Nova
running on bare metal and Cinder running in a container on the same
host.

The whole idea of AttachDB might need rethinking, but this should be
a quick hotfix.