spopenstack: allow no SP_OURID if .sync() is never called.
If the StorPool OpenStack integration is used on an iSCSI-only node,
there is no need for it to even have a SP_OURID setting in its
storpool.conf file - all that it does will be send API queries and
attach volumes via iSCSI to its IQN (done separately by the StorPool
Cinder driver). So do not error out if there is no SP_OURID value.

Change-Id: Ib72179e800f1d6e7c61a66b32a2d755c69135ec2